backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
pydantic-settings==2.5.0
groq==0.11.0
python-dotenv==1.0.1
python-multipart==0.0.12
aiosqlite==0.20.0
python-docx==1.1.2
